TMS-2279 provider/vagrant: rename vagrantkite_build -> vagrant_instance
