+++ go/src/koding/kites/cmd/provider-vagrant/main_test.go
