+++ go/src/koding/kites/kloud/kloudctl/vagrant.json
@@ -1,13 +1,9 @@
 {
     "resource": {
-        "vagrantkite_build": {
+        "vagrant_instance": {
             "example_1": {
-                "cpus": 2,
-                "memory": 2048,
                 "user_data": "sudo apt-get install sl -y\ntouch /tmp/${var.koding_user_username}.txt",
 
-                "registerURL": "http://example1.rafal.tunnel.dev.koding.io:8081/klient/kite",
-                "kontrolURL": "http://koding-rjeczalik.ngrok.com/kontrol/kite",
                 "klientURL": "https://s3-eu-west-1.amazonaws.com/kodingdev-klient/klient_0.1.99_development_amd64.deb"
             }
         }
