+++ go/src/koding/kites/terraformplugins/vagrant/README.md
@@ -1,4 +1,4 @@
-# terraform-provider-vagrantkite
+# terraform-provider-vagrant
 Vagrant Kite Provider for Terraform
 
 # Usage
@@ -26,6 +26,8 @@ Advanced optional fields (only use them if you know what you do):
   from the kite.key automatically and Klient registers to that Kontrol URl. If
   we pass this field, klient will automatically register itself to this URL
   instead of the from kite.key.
+- klientURL: Overwrite URL to get the klient.deb during provisioning of
+  the vagrant box.
 
 Computed fields:
 
@@ -35,7 +37,7 @@ Computed fields:
 # Example
 
 ```
-resource "vagrantkite_build" "myfirstvm" {
+resource "vagrant_instance" "myfirstvm" {
     filePath = "/home/etc/Vagrantfile"
     queryString = "///////8c396fd6-c91c-4454-45c2-5c461ad32645"
     provisionData = "eyJVc2VybmFtZSI6ImFyc2xhbiIsIkhvc3RuYW1lIjoiYXJzbGFuIiwiR3J"
@@ -45,7 +47,7 @@ resource "vagrantkite_build" "myfirstvm" {
 or for example change the `memory` or `cpus` fields:
 
 ```
-resource "vagrantkite_build" "myfirstvm" {
+resource "vagrant_instance" "myfirstvm" {
     filePath = "/home/arslan/myGoApp/Vagrantfile"
     queryString = "///////8c396fd6-c91c-4454-45c2-5c461ad32645"
     provisionData = "eyJVc2VybmFtZSI6ImFyc2xhbiIsIkhvc3RuYW1lIjoiYXJzbGFuIiwiR3J"
